#27 fix : 수정
