+++ src/main/java/ticketing/ticketing/Service/InquiryService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import ticketing.ticketing.DTO.InquiryRequestDto;
 import ticketing.ticketing.DTO.InquiryResponseDto;
@@ -34,38 +35,39 @@ public class InquiryService {
     private final InquiryFileRepository inquiryFileRepository;
     private final UserInquiryRepository userInquiryRepository;
 
-    private final String uploadPath = "uploads"; // 저장할 경로 지정
+    private final String uploadPath = "uploads";
 
-    // ✅ 전체 문의 목록 조회
-    public Page<InquiryResponseDto> getInquiries(Pageable pageable) {
-        return inquiryRepository.findAll(pageable)
+    // ✅ 사용자별 문의 목록 조회 (페이지네이션)
+    public Page<InquiryResponseDto> getInquiriesByUser(Long userId, Pageable pageable) {
+        return inquiryRepository.findByUserId(userId, pageable)
                 .map(InquiryResponseDto::fromEntity);
     }
 
-    // ✅ 단건 상세 조회
-    public InquiryResponseDto getInquiryById(Long id) {
-        Inquiry inquiry = inquiryRepository.findById(id)
-                .orElseThrow(() -> new InquiryNotFoundException("해당 ID의 문의가 존재하지 않습니다."));
+    // ✅ 단건 조회 (권한 체크: userId와 문의 소유자 일치 여부 확인)
+    public InquiryResponseDto getInquiryByIdAndUser(Long inquiryId, Long userId) {
+        Inquiry inquiry = inquiryRepository.findByIdAndUserId(inquiryId, userId)
+                .orElseThrow(() -> new InquiryNotFoundException("해당 ID의 문의가 존재하지 않거나 권한이 없습니다."));
         return InquiryResponseDto.fromEntity(inquiry);
     }
 
-    // ✅ 1:1 문의 등록
+    // ✅ 문의 생성 (파일 저장 포함, 트랜잭션 보장)
+    @Transactional(rollbackFor = Exception.class)
     public InquiryResponseDto createInquiryWithFiles(Long userId, InquiryRequestDto dto, List<MultipartFile> files) {
         if (files != null && files.size() > 5) {
             throw new IllegalArgumentException("최대 5개의 파일만 업로드 가능합니다.");
         }
 
-        // 사용자 조회
+        // 유저 확인
         User user = userInquiryRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException("사용자를 찾을 수 없습니다."));
 
-        // Inquiry 생성자 방식으로 객체 생성
+        // 문의 저장
         Inquiry inquiry = new Inquiry(
                 user,
                 dto.getTitle(),
                 dto.getContent(),
                 InquiryType.valueOf(dto.getType()),
-                InquiryStatus.PENDING, // enum 정의에 맞게 변경
+                InquiryStatus.PENDING,
                 LocalDateTime.now()
         );
         inquiryRepository.save(inquiry);
@@ -87,7 +89,16 @@ public class InquiryService {
         return InquiryResponseDto.fromEntity(inquiry);
     }
 
-    // ✅ 파일 유효성 검사
+    // ✅ 문의 완료 처리 (관리자용)
+    @Transactional
+    public void markInquiryAsCompleted(Long inquiryId) {
+        Inquiry inquiry = inquiryRepository.findById(inquiryId)
+                .orElseThrow(() -> new InquiryNotFoundException("문의가 존재하지 않습니다."));
+
+        inquiry.markCompleted(LocalDateTime.now());
+    }
+
+    // ✅ 파일 검증
     private void validateFile(MultipartFile file) {
         if (file.getSize() > 5 * 1024 * 1024) {
             throw new IllegalArgumentException("파일은 5MB를 초과할 수 없습니다.");
@@ -99,7 +110,7 @@ public class InquiryService {
         }
     }
 
-    // ✅ 파일 저장 처리
+    // ✅ 파일 저장
     private String saveFile(MultipartFile file) {
         String uuid = UUID.randomUUID().toString();
         String fileName = uuid + "_" + file.getOriginalFilename();
